Add request steps table to the flow page.
