+++ apps/backend/src/core/request/resolvers/requestStepResolver.ts
@@ -11,6 +11,7 @@ export const requestStepResolver = ({
 }: {
   reqStep: RequestStepPrismaType;
   step: StepPrismaType;
+  
 }): RequestStep => {
   const res: RequestStep = {
     requestStepId: reqStep.id,
