+++ apps/frontend/src/pages/Requests/Requests.tsx
@@ -1,101 +1,12 @@
 import Typography from "@mui/material/Typography";
 import PageContainer from "@/layout/PageContainer";
-import { GetRequestStepsDocument, RequestStepSummaryFragment } from "@/graphql/generated/graphql";
-import Box from "@mui/material/Box";
-import { useLazyQuery } from "@apollo/client";
-import { ChangeEvent, useEffect, useState } from "react";
-
-import CreateButton from "@/components/Menu/CreateButton";
-import Loading from "@/components/Loading";
-import Search from "@/components/Tables/Search";
-import { RequestStepsTable } from "./RequestsTable";
-import { Button, debounce } from "@mui/material";
+import { RequestStepsSearch } from "./RequestStepsSearch";
 
 export const Requests = () => {
-  const [searchQuery, setSearchQuery] = useState("");
-  const [oldCursor, setOldCursor] = useState<string | undefined>(undefined);
-  const limit = 20;
-  // const [statusToggle, setStatusToggle] = useState<"open" | "closed">("open");
-
-  const [getResults, { loading, data, fetchMore }] = useLazyQuery(GetRequestStepsDocument);
-
-  const debouncedRefetch = debounce(() => {
-    getResults({ variables: { userOnly: true, flowId: null, searchQuery, limit } });
-  }, 1000);
-
-  useEffect(() => {
-    debouncedRefetch();
-  }, [
-    searchQuery, // debouncedRefetch
-  ]);
-
-  useEffect(() => {
-    getResults({ variables: { userOnly: true, flowId: null, searchQuery, limit } });
-  }, []);
-
-  const newCursor = data?.getRequestSteps[data.getRequestSteps.length - 1].id;
-  const requestSteps = (data?.getRequestSteps ?? []) as RequestStepSummaryFragment[];
-
   return (
     <PageContainer>
       <Typography variant="h1">Requests</Typography>
-      <Box
-        sx={{
-          display: "flex",
-          flexDirection: "column",
-          gap: "30px",
-        }}
-      >
-        <Box
-          sx={{
-            width: "100%",
-            display: "flex",
-            justifyContent: "space-between",
-            flexDirection: "row",
-            gap: "16px",
-          }}
-        >
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "row",
-              gap: "16px",
-              width: "100%",
-              maxWidth: "500px",
-            }}
-          >
-            <Search
-              searchQuery={searchQuery}
-              changeHandler={(event: ChangeEvent<HTMLInputElement>) => {
-                setSearchQuery(event.target.value);
-              }}
-            />
-
-            {/* <StatusToggle status={statusToggle} setStatus={setStatusToggle} /> */}
-          </Box>
-          <CreateButton />
-        </Box>
-        {loading ? <Loading /> : <RequestStepsTable requestSteps={requestSteps} />}
-        {/* if there are no new results or no results at all, then hide the "load more" button */}
-        {oldCursor !== newCursor && (data?.getRequestSteps.length ?? 0) >= limit && (
-          <Button
-            onClick={() => {
-              setOldCursor(newCursor);
-              return fetchMore({
-                variables: {
-                  cursor: newCursor,
-                  userOnly: true,
-                  flowId: null,
-                  searchQuery,
-                  limit,
-                },
-              });
-            }}
-          >
-            Load more
-          </Button>
-        )}
-      </Box>
+      <RequestStepsSearch userOnly  />
     </PageContainer>
   );
 };
